requirements.txt
Django==3.1.3
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.2
dj-database-url==0.5.0
dj-static==0.0.6
pip==20.2.4
python-decouple==3.3
requests==2.24.0